+++ internal/command/parser_test.go
@@ -1,4 +1,4 @@
-package gokit
+package command
 
 import (
 	"go/ast"
